+++ resources/envs/main.tf
@@ -67,11 +67,6 @@ variable "aws_cf_cdn_arn" {
   description = "CloudFront distribution ARN for the BAS CDN production environment used for item thumbnails."
 }
 
-variable "aws_cf_site_replica_id" {
-  type        = string
-  description = "CloudFront distribution identifier for the UKRI parallel/replica live static site hosting."
-}
-
 locals {
   static_site_ref         = "v0.6.1"       # Static site module version.
   static_site_tls_version = "TLSv1.2_2025" # for compatibility with BAS load balancer
@@ -80,14 +75,14 @@ locals {
 
 provider "aws" {
   region = "eu-west-1"
-  # credentials set by awscli profile
+  # credentials set by awscli default profile
 }
 
 provider "aws" {
   # alias for resources requiring the 'us-east-1' region, which is used as a control region by AWS for some services.
   alias  = "us-east-1"
   region = "us-east-1"
-  # credentials set by awscli profile
+  # credentials set by awscli default profile
 }
 
 provider "cloudflare" {
@@ -549,11 +544,6 @@ output "sentry_dsn" {
 output "site_cf_id" {
   value       = module.site_prod.cloudfront_distribution_id
   description = "CloudFront distribution ID for live static site hosting."
-}
-
-output "site_replica_cf_id" {
-  value       = var.aws_cf_site_replica_id
-  description = "CloudFront distribution ID for UKRI parallel/replica live static site hosting."
 }
 
 output "thumbnails_cf_id" {